requirements.txt
aiogram
psycopg2-binary
dotenv

fastapi
requests
beautifulsoup4
python-dotenv
uvicorn
httpx
pydantic